+++ superset/assets/src/welcome/App.jsx
@@ -26,7 +26,7 @@ import messageToastReducer from '../messageToasts/reducers';
 import { initEnhancer } from '../reduxUtils';
 import setupApp from '../setup/setupApp';
 import Welcome from './Welcome';
-import Menu from '../components/Menu';
+import Menu from '../components/Menu/Menu';
 
 setupApp();
 
@@ -48,10 +48,10 @@ const store = createStore(
 
 const App = () => (
   <Provider store={store}>
-    <React.Fragment>
+    <>
       <Menu data={menu} />
       <Welcome user={user} />
-    </React.Fragment>
+    </>
   </Provider>
 );
